pyproject.toml
[project]
name = "homeser-backend"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
    "asgiref==3.9.1",
    "cloudinary==1.44.1",
    "dj-database-url==2.1.0",
    "django==5.0",
    "django-cachalot>=2.8.0",
    "django-cache-machine>=1.2.0",
    "django-cloudinary-storage==0.3.0",
    "django-cors-headers==4.3.1",
    "django-environ==0.11.2",
    "django-fsm>=3.0.0",
    "django-guardian>=3.2.0",
    "django-lifecycle>=1.2.4",
    "django-model-utils>=5.0.0",
    "django-ratelimit>=4.1.0",
    "django-redis==6.0.0",
    "djangorestframework==3.14.0",
    "django-filter>=24.2",
    "djangorestframework-simplejwt==5.3.0",
    "drf-extensions>=0.8.0",
    "drf-spectacular==0.27.2",
    "drf-writable-nested>=0.7.2",
    "gunicorn==21.2.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "packaging==25.0",
    "pillow==11.3.0",
    "psycopg2-binary==2.9.10",
    "pyjwt==2.10.1",
    "python-decouple>=3.8",  # Used in test files
    "python-dotenv>=1.1.1",  # Used in settings.py
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "redis==6.4.0",
    "referencing==0.36.2",
    "requests==2.32.3",
    "rpds-py==0.27.1",
    "setuptools>=80.9.0",  # Needed for pkg_resources
    "sqlparse==0.5.3",
    "sslcommerz-python-api==1.0.1",
    "textblob>=0.19.0",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3==2.3.0",
    "whitenoise==6.6.0",
    "pydantic>=2.11.9",
    "django-bulk-sync>=3.3.0",
    "temporalio>=1.18.0",
    "opentelemetry-instrumentation-django>=0.58b0",
    "sentry-sdk>=2.39.0",
]

[project.optional-dependencies]
dev = [
    "astroid==2.3.3",
    "autoflake>=2.3.1",
    "bandit>=1.8.6",
    "isort==4.3.21",
    "mccabe==0.6.1",
    "pyright>=1.1.405",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "radon>=6.0.1",
    "ruff>=0.12.12",
    "setuptools>=80.9.0",
    "vulture>=2.14",
    "wrapt==1.11.2",
    "factory-boy>=3.3.3",
]

[tool.uv]
package = false
